create a custom hook to fetch the data and manage state of loading
